fix: cast ALL loss inputs to fp32 for bf16 training compatibility
Root cause (from brainstorm analysis): Test notebook passed because it
used out.sum() as loss, completely bypassing CombinedLoss. Training
crashed because CombinedLoss receives bf16 tensors and some backward
ops (interpolate, sqrt, normalize) lack bf16 kernels.

Fix: Cast pred, img_feat, text_feat, pixel_feat, aux_preds to fp32
at CombinedLoss.forward() entry. All sub-losses (Dice, CE, Edge,
Contrastive, Deep Supervision) now run entirely in fp32.
PyTorch autograd handles fp32→bf16 gradient flow automatically.

Also: EdgeLoss Sobel buffers explicitly cast to fp32 (they follow
model.to(bf16) since they're registered buffers).

Re-enabled deep_supervision=true in V5.1 config.

Codex audit: 5 issues identified, all covered by this fix.

+++ losses/edge_loss.py
@@ -29,9 +29,10 @@ class EdgeLoss(nn.Module):
     def get_edge_mask(self, target: torch.Tensor) -> torch.Tensor:
         target_float = target.float().unsqueeze(1)
 
-        gx = F.conv3d(target_float, self.sobel_x, padding=1)
-        gy = F.conv3d(target_float, self.sobel_y, padding=1)
-        gz = F.conv3d(target_float, self.sobel_z, padding=1)
+        # Sobel kernels may be bf16 after model.to(bf16) — force fp32 for conv
+        gx = F.conv3d(target_float, self.sobel_x.float(), padding=1)
+        gy = F.conv3d(target_float, self.sobel_y.float(), padding=1)
+        gz = F.conv3d(target_float, self.sobel_z.float(), padding=1)
         edge = torch.sqrt(torch.clamp_min(gx**2 + gy**2 + gz**2, 0.0))
         edge = edge / torch.clamp_min(edge.amax(dim=(-3, -2, -1), keepdim=True), 1e-4)
         return edge
